+++ modules/application.js
@@ -101,8 +101,12 @@ define([
 
     // Routes configuration
     w20CoreApplication.config([ '$routeProvider', '$locationProvider', '$sceDelegateProvider', function ($routeProvider, $locationProvider, $sceDelegateProvider) {
-        // Use the classic hashbang prefix
         $locationProvider.hashPrefix('!');
+        if (config.prettyUrls) {
+            $locationProvider.html5Mode(true);
+        } else {
+            $locationProvider.html5Mode(false);
+        }
 
         // WhiteList all trusted URLs
         $sceDelegateProvider.resourceUrlWhitelist($sceDelegateProvider.resourceUrlWhitelist().concat(sceUrlWhiteList));
@@ -139,7 +143,7 @@ define([
             };
 
             if (typeof allRouteHandlers[route.type] === 'undefined') {
-                console.warn('no handler for route type ' + route.type + ' when registering ' + route.path);
+                console.warn('No handler for route type ' + route.type + ' when registering ' + route.path);
             } else {
                 route = allRouteHandlers[route.type](route);
             }
@@ -407,7 +411,7 @@ define([
             run: function (modules, fragments, callback) {
                 var angularModules = [],
                     bootstrap = function () {
-                        w20.console.log('bootstrapping angular application with the following modules: ' + angularModules);
+                        w20.console.log('Bootstrapping angular application with the following modules: ' + angularModules);
 
                         angular.bootstrap(window.document, angularModules, {
                             strictDi: !w20.debug