+++ test-main.js
@@ -37,6 +37,12 @@ window.w20 = {
             vars: {
                 'components-path': '/base/bower_components'
             }
+        },
+        '/base/optional-fragment.w20.json': {
+            optional: true
+        },
+        '/base/ignored-fragment.w20.json': {
+            ignore: true
         }
     },
     deps: tests,