+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 2.2.3 (?)
+
+* [new] Add support for HTML5 mode (pretty urls)
+* [new] Add the ability to specify an `optional` attribute on any fragment configuration, allowing application to load anyway.
+* [new] Add the ability to specify an `ignore` attribute on any fragment configuration to avoid loading it.
+
 # Version 2.2.2 (2016-02-15)
 
 * [fix] Remove usage of non-standard `trimLeft` and `trimRight` functions in `w20.js`